UNet/requirements.txt
opencv-python          == 4.7.0.72
opencv-python-headless == 4.7.0.72
pandas                 == 1.5.3
Pillow                 == 9.4.0
numpy                  == 1.24.2
matplotlib             == 3.7.0
torch                  == 1.13.1
torchvision            == 0.14.1
scikit-learn           == 1.2.1